Refactor qrcode support
